Head back to the template. Everything about the form is controlled with these HTML attributes. Obviously, the most important one is amount. Set it to {{ cart.total }} - cart is a variable I've passed into the template - then the important part: * 100.
Whenever you talk about amounts in Stripe, you use the smallest denomination of the currency, so cents in USD. If you need to charge the user , then tell Stripe to charge them an amount of 500 cents.

Then, fill in anything else that's important to you, for example, data-image. I'll set this to our logo.

Checking out with a Test Card

Refresh to reflect the new settings. The total should be 2, and it is! Because we're using the test environment, Stripe gives us fake, test cards we can use to checkout. I'll show you others later - but to checkout successfully, use 4242 4242 4242 4242. You can use any valid future expiration and any CVC.

Ok, moment of truth: hit pay!

It worked! I think... Wait, what just happened? Well, a really important step just happened - a step that's core to how Stripe checkout works.

The Stripe Checkout Token

First, credit card information is never sent to our servers... which is the greatest news ever from a security standpoint. I do not want to handle your CC number: this would greatly increase the security requirements on my server.

Instead, when you hit 'Pay', this sends the credit card information to Stripe directly, via AJAX. If the card is valid, it sends back a token string, which represents that card. The Stripe JS puts that token into the form as an hidden input field and then submits the form like normal to our server. So the only thing that's sent to our server is this token. The customer has not been charged yet, but with a little more work - we can fetch that token in our code and ask Stripe to charge that credit card.