To actually charge a user, we need to... well, create a Stripe charge In the Stripe API, click 'Charges' on the left and find Create a Charge.
Hey! It wrote the code for us. Copy the code block on the right. Now, go back to OrderController and first, create a new {dollar_sign}token variable and set it to the stripeToken POST parameter.

My editor now sees all these new Stripe classes.

See that API key?

Once again, this is a real key from our account in the test environment. This time, it's the secret key. The public key is the one in our template.

Update charge details with our real information. To get the total, I'll fetch a service I created for the project called shopping_cart, call getTotal() on this, and then multiply it by 100.

For source, replace this fake token with the submitted {dollar_sign}token variable.

The token basically represents the credit card that was just sent. We're saying: Use this card as the source for this charge. And then, put whatever you want for description, like 'First test charge'

When this code runs, it will make an API request to Stripe. If that's successful, the user will be charged. If something goes wrong, Stripe will throw an Exception. More on that later.