We already know that our Stripe account has two environments, and each has its own two keys. This means that when we deploy, we'll need to update our code to use these Live keys, instead of the ones from the Test environment.
Well... that's going to be a bummer: the public key is hard-coded right in the middle of my template:
53 lines  app/Resources/views/order/checkout.html.twig
... lines 1 - 3
{% block body %}
<div class='nav-space-checkout'>
  <div class='container'>
    <div class='row'>
... lines 8 - 34
      <div class='col-xs-12 col-sm-6'>
        <form action='' method='POST'>
          <script
... line 38
            data-key='pk_test_HxZzNHy8LImKK9LDtgMDRBwd'
... lines 40 - 45
          </script>
        </form>
      </div>
    </div>
  </div>
</div>
{% endblock %}

and the private one is stuck in the center of a controller:

58 lines  src/AppBundle/Controller/OrderController.php
... lines 1 - 10
class OrderController extends BaseController
{
... lines 13 - 30
  public function checkoutAction(Request {dollar_sign}request){
... lines 33 - 34
    if ({dollar_sign}request->isMethod('POST')) {
... lines 36 - 37
      \Stripe\Stripe::setApiKey('XXX_PRIVATEKEY_XXX');
... lines 39 - 49
    }
... lines 51 - 56
  }
}

If you love editing random files whenever you deploy, then this is perfect! Have fun!

But for the rest of us, we need to move these keys to a central configuration file so they're easy to update on deploy. We also need to make sure that we don't commit this private key to our Git repository... ya know... because it's private - even though I keep showing you mine.