How you do this next step will vary for different frameworks, but is philosophically always the same. In Symfony, we're going to move our keys to a special parameters.yml file, because our project is setup to not commit this to Git.
Add a stripe_secret_key config and set its value to the key from the controller:

 22 lines  app/config/parameters.yml.dist
... lines 1 - 3
parameters:
... lines 5 - 19
    stripe_secret_key: XXX_PRIVATEKEY_XXX
... lines 21 - 22
Then add stripe_public_key and set that to the one from the template:

 22 lines  app/config/parameters.yml.dist
... lines 1 - 3
parameters:
... lines 5 - 20
    stripe_public_key: YYY_PUBLISHABLE_KEY_YYY
In Symfony, we also maintain this other file - parameters.yml.dist - as a template for the original, uncommitted file. This one is committed to the repository. Add the keys here too, but give them fake values.