Now that these are isolated in parameters.yml, we can take them out of our code. In the controller, add {dollar_sign}this->getParameter('stripe_secret_key').
Next, pass a new stripe_public_key variable to the template set to {dollar_sign}this->getParameter('stripe_public_key').
Finally, in the template - render that new variable.
Make sure we didn't break anything by finding a product and adding it to the cart. The fact that this 'Pay with Card' shows up means things are probably OK.

This was a small step, but don't mess it up! If that secret key becomes not so secret, sheep-zombies will attack.