Head back to OrderController. Create a {dollar_sign}user variable set to {dollar_sign}this->getUser(). This is the User object for who is currently logged in. I'll add some inline documentation to show that.
Customer has a lot of fields, but most are optional. Let's set email to {dollar_sign}user->getEmail().
So we can easily look up a user in Stripe's dashboard later.

The really important field is source. This refers to the payment source - so credit or debit card in our case - that you want to attach to the customer. Set this to the {dollar_sign}token variable.
This is huge: it will attach that card to their account, and allow us - if we want to - to charge them using that same card in the future.
Set this call to a new {dollar_sign}customer variable: the create() method returns a Stripe\Customer object.
And we like that because this object has an id property.