One InvoiceItem per Product
This now gives us more flexibility. Since sheep love to shop, they'll often buy multiple products. In fact, let's go buy some shears, and some Sheerly Conditioned. If we checked out right now, this would show up as one giant line item for {dollar_sign}106.00 on the Invoice. We can do better than that.

In OrderController, around the InvoiceItem part, add a foreach, over {dollar_sign}this->get('shopping_cart')->getProducts() as {dollar_sign}product. In other words, let's loop over all the actual products in our cart and create a separate InvoiceItem for each. All we need to do is change the amount to be: {dollar_sign}product->getPrice() * 100. We can even improve the description: set it to {dollar_sign}product->getName()

Now, if we eventually send the user a receipt, it's going to be very easy to look on this Stripe invoice and see exactly what we charged them for.

Test time! Put in our awesome fake information, hit ENTER... and no errors.

In Stripe, click back to the customer page and find the new invoice - for {dollar_sign}106. Click on that. Yes! 2 crystal clear invoice line items.

So yes, you can just charge customers. But if you create an Invoice with detailed line items, you're going to have a much better accounting system in the long run.