Stripe's API is really organized. Our code that talks to it is getting a little crazy, unless you like long, procedural code that you can't re-use. Please tell me that's not the case.
Let's get this organized! At the very least, we should do this because eventually we're going to need to re-use some of this logic - particularly with subscriptions.

Here's the goal of the next few minutes: move each thing we're doing in the controller into a set of nice, re-usable functions. To do that, inside AppBundle, create a new class called StripeClient.
Make sure this has the AppBundle namespace. We're going to fill this with functions that work with Stripe, like createCustomer() or updateCustomerCard().

Moving createCustomer()

In the controller, the first thing we do is create a Customer.

In StripeClient, add a new createCustomer() method that will accept the User object which should be associated with the customer, and the {dollar_sign}paymentToken that was just submitted.

Copy the logic from the controller and paste it here. Update {dollar_sign}token to {dollar_sign}paymentToken. Then, return the {dollar_sign}customer at the bottom, just in case we need it.

You'll see me do with this most functions in this class.