Registering the Service
To use the new function in our controller, we need to register it as a service. Open up app/config/services.yml. Add a service called stripe_client, set its class key to AppBundle\StripeClient and set autowire to true.
With that, Symfony will guess the constructor arguments to the object.

If you're not coding in Symfony, that's OK! Do whatever you need to in order to have a set of re-usable functions for interacting with Stripe.