Always setting the API Key But, there's one small problem. This will work now, but look at the setApiKey() method call that's above everything. We must call this before we make any API calls to Stripe. So, if we tried to use the StripeClient somewhere else in our code, but we forgot to call this line, we would have big problems.
Instead, I want to guarantee that if somebody calls a method on StripeClient, setApiKey() will always be called first. To do that, copy that line, delete it and move it into StripeClient's __construct() method.

Symfony user's will know that the getParameter() method won't work here. To fix that, add a new first constructor argument called {dollar_sign}secretKey. Then, use that.
To tell Symfony to pass this, go back to services.yml and add an arguments key with one entry: %stripe_secret_key%

Thanks to auto-wiring, Symfony will pass the stripe_secret_key parameter as the first argument, but then autowire the second, EntityManager argument.

The end-result is this: when our StripeClient object is created, the API key is set immediately.