Wait, life is not good. I hate needing to setup SSL certificates on my local machine. I actually have one setup already, but other developers might not. That's a huge pain for them... for no benefit.
Fortunately, there's a trick. Replace https, with %secure_channel%

This syntax is referencing a parameter in Symfony, so basically a configuration variable. Open parameters.yml, add a new secure_channel parameter and set it to http:
25 lines  app/config/parameters.yml.dist
... lines 1 - 3
parameters:
... lines 5 - 22
    # set to https on production
    secure_channel: http
And as you know, if you add a key here, also add it to parameters.yml.dist:

 25 lines  app/config/parameters.yml.dist
... lines 1 - 3
parameters:
... lines 5 - 22
    # set to https on production
    secure_channel: http
Ok, head back to the homepage: http://localhost:8000 and click to checkout. Hey! We're back in http. When you deploy, change that setting to https and boom, your checkout will be secure.

So there's your little trick for forcing https without being forced to hate your life while coding.