Now that we've got the form in place, we need to add some JavaScript that will send all the card information to Stripe. Once again, Stripe wrote a lot of this code for us. Over-achiever.
Copy their JavaScript and scroll up and paste it with our other JavaScript.

Then, back on the docs, scroll down a little further to another function called stripeResponseHandler(). Copy that too and paste it.

Let's look at the code: it uses a jQuery document.ready block to find the form and attach an on submit handler function:

 89 lines  app/Resources/views/order/checkout.html.twig
... lines 1 - 3
{% block javascripts %}
... lines 5 - 8
  <script type='text/javascript'>
... lines 10 - 11
    {dollar_sign}(function () {
      var {dollar_sign}form = {dollar_sign}('#payment-form');
      {dollar_sign}form.submit(function (event) {
... lines 15 - 22
      });
    });
... lines 25 - 47
  </script>
{% endblock %}
... lines 50 - 89
Because basically, when the user submits the form, we don't want to submit the form! Ahem, I mean, we want to stop, send all the information to Stripe, wait for the token to come back, put that in the form, and then submit it.