Finally, here is the meat of the code. When we call Stripe.card.createToken():
 92 lines  app/Resources/views/order/checkout.html.twig
... lines 1 - 3
{% block javascripts %}
... lines 5 - 8
  <script type='text/javascript'>
... lines 10 - 11
    {dollar_sign}(function () {
... line 13
      {dollar_sign}form.submit(function (event) {
... lines 15 - 19
        // Request a token from Stripe:
        Stripe.card.createToken({dollar_sign}form, stripeResponseHandler);
      });
    });
... lines 24 - 50
  </script>
{% endblock %}
... lines 53 - 92
Stripe's Javascript will automatically fetch all the credit card data by reading the data-stripe attributes. Then, it sends those to stripe via AJAX. When that call finishes, it will execute the stripeResponseHandler function, and hopefully the response will contain that all-important token:

 92 lines  app/Resources/views/order/checkout.html.twig
... lines 1 - 3
{% block javascripts %}
... lines 5 - 8
  <script type='text/javascript'>
... lines 10 - 24
    function stripeResponseHandler(status, response) {
... lines 26 - 49
    }
  </script>
{% endblock %}
... lines 53 - 92
Now, if there was a problem with that card - like an invalid expiration - we need to show that error to the user. To do that, it looks for a payment-errors class and puts the message inside of that.

92 lines  app/Resources/views/order/checkout.html.twig
... lines 1 - 3
{% block javascripts %}
... lines 5 - 8
  <script type='text/javascript'>
... lines 10 - 24
    function stripeResponseHandler(status, response) {
... lines 26 - 28
      if (response.error) { // Problem!
        // Show the errors on the form:
        {dollar_sign}form.find('.js-checkout-error')
          .text(response.error.message)
... lines 34 - 48
      }
    }
  </script>
{% endblock %}
... lines 53 - 92
I have a div ready for this. Its class is js-checkout-error and its hidden by default:

 68 lines  app/Resources/views/order/_cardForm.html.twig
<form action='' method='POST' class='js-checkout-form checkout-form'>
... lines 2 - 53
  <div class='row'>
    <div class='col-xs-8 col-sm-6 col-sm-offset-2 text-center'>
      <div class='alert alert-danger js-checkout-error hidden'></div>
    </div>
  </div>
... lines 59 - 66
</form>

Change the selector to .js-checkout-error, set the text, but then also call removeClass('hidden') so the element shows up.

Below in the else, life is good! I'll paste the .js-checkout-error code from before and modify it to re-add the hidden class - since now things are successful.

When things work, the response comes back with a token, which we get via response.id:
92 lines  app/Resources/views/order/checkout.html.twig
... lines 1 - 3
{% block javascripts %}
... lines 5 - 8
  <script type='text/javascript'>
... lines 10 - 24
    function stripeResponseHandler(status, response) {
... lines 26 - 28
      if (response.error) { // Problem!
... lines 30 - 36
      } else { // Token was created!
... lines 38 - 40
        // Get the token ID:
        var token = response.id;
... lines 43 - 48
      }
    }
  </script>
{% endblock %}
... lines 53 - 92
To send this to the server, we just smash it into a new input hidden field called... drumroll ... stripeToken:

 92 lines  app/Resources/views/order/checkout.html.twig
... lines 1 - 3
{% block javascripts %}
... lines 5 - 8
  <script type='text/javascript'>
... lines 10 - 24
    function stripeResponseHandler(status, response) {
... lines 26 - 28
      if (response.error) { // Problem!
... lines 30 - 36
      } else { // Token was created!
... lines 38 - 43
        // Insert the token ID into the form so it gets submitted to the server:
        {dollar_sign}form.append(.val(token));
... lines 46 - 48
      }
    }
  </script>
{% endblock %}
... lines 53 - 92

To send this to the server, we just smash it into a new input hidden field called... drumroll ... stripeToken. This is precisely what the embedded form did. Once the form is submitted, the controller will hum along like nothing ever changed.

Testing the Error and Success

But, that's assuming we didn't mess something up! That's a big but. Go back and refresh the page.

First, test that the error handling works by adding an expiration date in the past. Put in the real credit card number -- oof, ugly formatting - we'll fix that. Then, use an expired expiration. Hit checkout and... boom!

It sent the info to Stripe, Stripe came back with an error, we put the error in the box, and showed that box to our user. In other words, we're awesome. Change this to a future expiration and try again.

It's alive!

The only problem I can think of now is how ugly entering a credit card number is: all those numbers just run together. The expiration field is a mess too. Oof. Let's fix that - it's surprisingly easy!