The one error we need to handle is card_error - because this happens when a card is declined.
To do that, let's move all of this processing logic into its own private function in this class. That'll make things cleaner.

To do this, I'll use a PhpStorm shortcut: select the code, hit Control+T (or go to the 'Refactor'->'Refactor This' menu) and select 'Method'. Create a new method called chargeCustomer(). Hit refactor. You don't need PhpStorm to do that: it just moved my code down into this private function and called that function from the original spot.

Handling the Card Exception

OK, back to business: we know that when a card is declined, something in that code will throw a Stripe\Error\Card exception. I'm adding a little documentation just to indicate this.
Back in checkoutAction(), add a new {dollar_sign}error = false variable before the if, because at this point, no error has occurred.
Next, surround the chargeCustomer() call in a try-catch: try chargeCustomer() and then catch just a \Stripe\Error\Card exception.
If we get here, there was a problem charging the card. Update  to some nice message, like: 'There was a problem charging your card.'. Then add {dollar_sign}e->getMessage(). That's will be the message that Stripe's sending back like, 'Your card was declined,' or, 'Your card cannot be used for this type of transaction.'