Now, if there is an error, we don't want to empty the cart, we don't want to add the nice message and we don't want to redirect. So, if (! {dollar_sign}error), then it's safe to do those things. If there is an error, our code will continue down and it will re-render the checkout template, which is exactly what we want! Pass in the {dollar_sign}error variable so we can show it to the user. Then, in then template, specifically the _cardForm template, render error inside of our error div. If there is no error, no problem! It won't render anything. If there is an error, then we we need to not render the hidden class. Use an inline if statement to say:
If error, then don't render any class, else render the hidden class
A little tricky, but that should do it.

Ok, let's try it again. Refresh the page. Put in the fake credit card: the number 4, a thousand zeroes, and a 2 at the end. Finish it up and submit.

There's the error! Setup, complete.

Ok, guys, you have a killer checkout system via Stripe. In part 2 of this course, we're going to talk about where things get a little bit more difficult: like subscriptions and discounts. This includes handling web hooks, one of the scariest and toughest parts of subscriptions.

But, don't stop - go make a great product and sell it.